refactor: move hooks into different files and remove unused hooks
